backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
python-dotenv
tensorflow
scikit-learn
